Fixing rate limiting

+++ routes/customer.php
@@ -24,13 +24,13 @@ Route::prefix('my')
     Route::get('notifications', [HomeController::class, 'notifications'])->name('notifications');
     Route::get('notifications/{notification}/open', [HomeController::class, 'openNotification'])->name('notifications.open');
     Route::get('bookings/{booking}/review',        [CustomerReviewController::class, 'create'])->name('reviews.create');
-    Route::post('bookings/{booking}/review',       [CustomerReviewController::class, 'store'])->name('reviews.store');
+    Route::post('bookings/{booking}/review',       [CustomerReviewController::class, 'store'])->middleware('throttle:review-write')->name('reviews.store');
     Route::get('bookings/{booking}/reschedule',    [BookingController::class, 'edit'])->name('bookings.reschedule');
     Route::patch('bookings/{booking}/reschedule',  [BookingController::class, 'update'])->name('bookings.reschedule.update');
 
     Route::get('reviews',             [CustomerReviewController::class, 'index'])->name('reviews.index');
     Route::get('reviews/{review}/edit', [CustomerReviewController::class, 'edit'])->name('reviews.edit');
-    Route::put('reviews/{review}',    [CustomerReviewController::class, 'update'])->name('reviews.update');
+    Route::put('reviews/{review}',    [CustomerReviewController::class, 'update'])->middleware('throttle:review-write')->name('reviews.update');
     Route::delete('reviews/{review}', [CustomerReviewController::class, 'destroy'])->name('reviews.destroy');
 
     Route::get('payments', [CustomerPaymentController::class, 'index'])->name('payments.index');
@@ -43,10 +43,10 @@ Route::prefix('my')
 
     Route::get('profile',          [ProfileController::class, 'edit'])->name('profile.edit');
     Route::put('profile',          [ProfileController::class, 'update'])->name('profile.update');
-    Route::put('profile/password', [ProfileController::class, 'updatePassword'])->name('profile.password');
-    Route::delete('profile',       [ProfileController::class, 'deactivate'])->name('profile.deactivate');
+    Route::put('profile/password', [ProfileController::class, 'updatePassword'])->middleware('throttle:password-confirm')->name('profile.password');
+    Route::delete('profile',       [ProfileController::class, 'deactivate'])->middleware('throttle:password-confirm')->name('profile.deactivate');
 
-    Route::put('profile/2fa',                 [ProfileController::class, 'toggleTwoFactor'])->name('profile.2fa.toggle');
+    Route::put('profile/2fa',                 [ProfileController::class, 'toggleTwoFactor'])->middleware('throttle:password-confirm')->name('profile.2fa.toggle');
     Route::delete('profile/devices/{device}', [ProfileController::class, 'removeTrustedDevice'])->name('profile.devices.destroy');
 
     Route::put('profile/email-notifications', [ProfileController::class, 'toggleEmailNotifications'])->name('profile.email-notifications');
